divide fragment(motcci'S changed at noon)
