+++ app/src/main/java/jp/co/cyberagent/dojo2020/MemoListFragment.kt
@@ -1,0 +1,13 @@
+package jp.co.cyberagent.dojo2020
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+
+class MemoListFragment: Fragment(){
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
+        return inflater.inflate(R.layout.memo_list_tab,container,false)
+    }
+}
