+++ app/src/main/java/jp/co/cyberagent/dojo2020/ProfileFragment.kt
@@ -1,0 +1,13 @@
+package jp.co.cyberagent.dojo2020
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+
+class ProfileFragment: Fragment(){
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
+        return inflater.inflate(R.layout.profile_tab,container,false)
+    }
+}
