--- app/src/main/java/jp/co/cyberagent/dojo2020/TabFragments.kt
@@ -1,24 +1,0 @@
-package jp.co.cyberagent.dojo2020
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-
-class EditRecordTabFragment: Fragment(){
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        return inflater.inflate(R.layout.edit_record_tab,container,false)
-    }
-}
-
-class TimerFragment: Fragment(){
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        return inflater.inflate(R.layout.timer_tab,container,false)
-    }
-}
-class ProfileFragment: Fragment(){
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        return inflater.inflate(R.layout.profile_tab,container,false)
-    }
-}
