+++ app/src/main/res/layout/memo_list_tab.xml
