+++ app/src/main/res/navigation/navigation_graph.xml
@@ -2,17 +2,17 @@
 <navigation xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:id="@+id/navigation_graph"
-    app:startDestination="@id/navi_smile">
+    app:startDestination="@id/navi_memo_list">
 
-    <fragment android:id="@+id/navi_smile"
-        android:name="jp.co.cyberagent.dojo2020.EditRecordTabFragment"
+    <fragment android:id="@+id/navi_memo_list"
+        android:name="jp.co.cyberagent.dojo2020.MemoListFragment"
         android:label="Fragment1"/>
 
-    <fragment android:id="@+id/navi_pets"
+    <fragment android:id="@+id/navi_timer"
         android:name="jp.co.cyberagent.dojo2020.TimerFragment"
         android:label="Fragment2"/>
 
-    <fragment android:id="@+id/navi_sun"
+    <fragment android:id="@+id/navi_profile"
         android:name="jp.co.cyberagent.dojo2020.ProfileFragment"
         android:label="Fragment3"/>
